+++ src/Field/Gallery.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Acf codifier gallery class
+ * ACF Codifier gallery class
  */
 
 namespace Geniem\ACF\Field;
@@ -8,7 +8,10 @@ namespace Geniem\ACF\Field;
 /**
  * Class Gallery
  */
-class Gallery extends Image {
+class Gallery extends \Geniem\ACF\Field {
+
+    use Common\Image;
+
     /**
      * Field type
      *
@@ -70,39 +73,5 @@ class Gallery extends Image {
      */
     public function get_max() {
         return $this->max;
-    }
-
-    /**
-     * Register a attachment validating function for the field
-     *
-     * @param callable $function A function to register.
-     * @return self
-     */
-    public function validate_attachment( $function ) {
-        $this->filters['validate_attachment'] = [
-            'filter'        => 'acf/validate_attachment/key=',
-            'function'      => $function,
-            'priority'      => 10,
-            'accepted_args' => 4,
-        ];
-
-        return $this;
-    }
-
-    /**
-     * Register a upload prefiltering function for the field
-     *
-     * @param callable $function A function to register.
-     * @return self
-     */
-    public function upload_prefilter( $function ) {
-        $this->filters['upload_prefilter'] = [
-            'filter'        => 'acf/upload_prefilter/key=',
-            'function'      => $function,
-            'priority'      => 10,
-            'accepted_args' => 3,
-        ];
-
-        return $this;
     }
 }