+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.15.0-beta] - 2018-11-20
+
+### Added
+- Wrapper file extensions methods on top of ACF's misleadingly named mime types.
+
+### Changed
+- Moved redundant functionalities to traits
+
 ## [1.14.4] - 2018-10-25
 
 ### Fixed